requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic-settings==2.10.1
numpy==2.2.4
Pillow==11.2.1
scikit-learn==1.7.0
python-dateutil==2.9.0.post0
rapidfuzz==3.13.0
pytesseract==0.3.13
opencv-python-headless==4.11.0.86
pdfplumber==0.11.7
python-docx==1.2.0
spacy==3.8.7
transformers==4.53.2
torch==2.7.1
redis==6.1.0
psycopg[binary]==3.2.9
SQLAlchemy==2.0.41
pytest==8.4.1
httpx==0.28.1
